moving around routes for new landing pages and adding more info button to main brands
