+++ src/pages/brands/bedtech.jsx
