+++ src/pages/brands/sealy/landing.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Layout from "../../../components/layout";
+
+const Landing = () => (
+  <Layout>
+    <div>
+      <h1>Sealy Landing Page</h1>
+    </div>
+  </Layout>
+);
+
+export default Landing;
