+++ src/pages/brands/stearns/landing.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Layout from "../../../components/layout";
+
+const Landing = () => (
+  <Layout>
+    <div>
+      <h1>Stearns Landing Page</h1>
+    </div>
+  </Layout>
+);
+
+export default Landing;
