+++ src/pages/brands/tempurpedic/landing.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Layout from "../../../components/layout";
+
+const Landing = () => (
+  <Layout>
+    <div>
+      <h1>Tempurpedic Landing Page</h1>
+    </div>
+  </Layout>
+);
+
+export default Landing;
